+++ crates/bevy_pbr/src/light_probe/environment_map.rs
@@ -44,13 +44,10 @@
 //!
 //! [several pre-filtered environment maps]: https://github.com/KhronosGroup/glTF-Sample-Environments
 
-use bevy_asset::{AssetId, Handle};
-use bevy_ecs::{
-    component::Component, query::QueryItem, reflect::ReflectComponent, system::lifetimeless::Read,
-};
+use bevy_asset::AssetId;
+use bevy_ecs::{query::QueryItem, system::lifetimeless::Read};
 use bevy_image::Image;
-use bevy_math::Quat;
-use bevy_reflect::{std_traits::ReflectDefault, Reflect};
+use bevy_light::EnvironmentMapLight;
 use bevy_render::{
     extract_instances::ExtractInstance,
     render_asset::RenderAssets,
@@ -71,56 +68,6 @@ use crate::{
 };
 
 use super::{LightProbeComponent, RenderViewLightProbes};
-
-/// A pair of cubemap textures that represent the surroundings of a specific
-/// area in space.
-///
-/// See [`crate::environment_map`] for detailed information.
-#[derive(Clone, Component, Reflect)]
-#[reflect(Component, Default, Clone)]
-pub struct EnvironmentMapLight {
-    /// The blurry image that represents diffuse radiance surrounding a region.
-    pub diffuse_map: Handle<Image>,
-
-    /// The typically-sharper, mipmapped image that represents specular radiance
-    /// surrounding a region.
-    pub specular_map: Handle<Image>,
-
-    /// Scale factor applied to the diffuse and specular light generated by this component.
-    ///
-    /// After applying this multiplier, the resulting values should
-    /// be in units of [cd/m^2](https://en.wikipedia.org/wiki/Candela_per_square_metre).
-    ///
-    /// See also <https://google.github.io/filament/Filament.html#lighting/imagebasedlights/iblunit>.
-    pub intensity: f32,
-
-    /// World space rotation applied to the environment light cubemaps.
-    /// This is useful for users who require a different axis, such as the Z-axis, to serve
-    /// as the vertical axis.
-    pub rotation: Quat,
-
-    /// Whether the light from this environment map contributes diffuse lighting
-    /// to meshes with lightmaps.
-    ///
-    /// Set this to false if your lightmap baking tool bakes the diffuse light
-    /// from this environment light into the lightmaps in order to avoid
-    /// counting the radiance from this environment map twice.
-    ///
-    /// By default, this is set to true.
-    pub affects_lightmapped_mesh_diffuse: bool,
-}
-
-impl Default for EnvironmentMapLight {
-    fn default() -> Self {
-        EnvironmentMapLight {
-            diffuse_map: Handle::default(),
-            specular_map: Handle::default(),
-            intensity: 0.0,
-            rotation: Quat::IDENTITY,
-            affects_lightmapped_mesh_diffuse: true,
-        }
-    }
-}
 
 /// Like [`EnvironmentMapLight`], but contains asset IDs instead of handles.
 ///
